Scripts to rename INPC to Property
