+++ src/Assets/ugui-mvvm/Editor/INPCBindingToPropertyBindingUpgradeEditor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3ba84edbb5772c4580f8d2a95a025f8
+guid: 51f28aaa640907c4ca06fb4d5cf96416
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
