+++ src/Assets/ugui-mvvm/PropertyBinding.cs.meta
@@ -1,10 +1,12 @@
 fileFormatVersion: 2
-guid: e3ba84edbb5772c4580f8d2a95a025f8
+guid: e3ee59b7cf539ba4db21f88739be3aeb
+timeCreated: 1435700164
+licenseType: Free
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 298
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
