+++ src/Assets/ugui-mvvm/PropertyExtensions/GameObjectPropertyExtensions.cs
@@ -3,10 +3,10 @@
 namespace uguimvvm.extensions
 {
     /// <summary>
-    /// Use this monobehavior to use INPC bindings to perform operations on gameobjects
+    /// Use this monobehavior to use PropertyBindings to perform operations on gameobjects
     /// e.g setting gameobject active state bound to a Viewmodel property
     /// </summary>
-    [RequireComponent(typeof(INPCBinding))]
+    [RequireComponent(typeof(PropertyBinding))]
     public class GameObjectPropertyExtensions : MonoBehaviour
     {
         public bool ActiveState
